requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
python-multipart==0.0.24
pdfplumber
python-dotenv==1.2.2
requests==2.33.1
streamlit==1.56.0
slowapi==0.1.9
